Added tests, added loggin inn UserController

+++ stock-market/Controllers/UserController.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Internal;
+using Microsoft.Extensions.Logging;
 using stock_market.DAL;
 using stock_market.Model;
 
@@ -14,15 +15,18 @@ namespace stock_market.Controllers
     public class UserController : ControllerBase
     {
         private readonly IUserRepository _db;
+        private readonly ILogger<UserController> _log;
 
-        public UserController(IUserRepository db)
+        public UserController(IUserRepository db, ILogger<UserController> log)
         {
             _db = db;
+            _log = log;
         }
 
 
         public async Task<string> GetFullName(int userid)
         {
+            _log.LogInformation("UserController: User got Fullname");
             return await _db.GetFullName(userid);
         }
 
@@ -42,9 +46,21 @@ namespace stock_market.Controllers
 
         }
 
-        public async Task<bool> CreateUser (User innUser)
+        public async Task<ActionResult> CreateUser (User innUser)
         {
-            return await _db.CreateUser(innUser);
+            if (ModelState.IsValid)
+            {
+                bool ok = await _db.CreateUser(innUser);
+                if (!ok)
+                {
+                    _log.LogError("UserController: Could not create user");
+                    return BadRequest("Could not create user");
+                }
+                _log.LogInformation("UserController: User created user");
+                return Ok("User created user");
+            }
+            _log.LogError("UserController: Fault in InputVal ");
+            return BadRequest("Fault in InputVal");
         }
 
         public async Task<List<User>> GetAll()
@@ -52,19 +68,56 @@ namespace stock_market.Controllers
             return await _db.GetAll();
         }
 
-        public async Task<bool> DeleteUser(int id)
+        public async Task<ActionResult> DeleteUser(int id)
         {
-            return await _db.DeleteUser(id);
+            if (ModelState.IsValid)
+            {
+                bool ok = await _db.DeleteUser(id);
+                if (!ok)
+                {
+                    _log.LogError("UserController: Could not delete user");
+                    return BadRequest("Could not delete user");
+                }
+                _log.LogInformation("UserController: User deleted user");
+                return Ok("User deleted user");
+            }
+            _log.LogError("UserController: Fault in InputVal ");
+            return BadRequest("Fault in InputVal");
         }
 
-        public async Task<bool> EditUser(User editUser) //Edits User, get current vaulues with GET and send new ones with POST. This is meant for the user to be able to change his own name.
+        public async Task<ActionResult> EditUser(User editUser) //Edits User, get current vaulues with GET and send new ones with POST. This is meant for the user to be able to change his own name.
         {
-            return await _db.EditUser(editUser);
+
+            if (ModelState.IsValid)
+            {
+                bool ok = await _db.EditUser(editUser);
+                if (!ok)
+                {
+                    _log.LogError("UserController: Could not edit user");
+                    return BadRequest("Could not edit user");
+                }
+                _log.LogInformation("UserController: User edit user");
+                return Ok("User edit user");
+            }
+            _log.LogError("UserController: Fault in InputVal ");
+            return BadRequest("Fault in InputVal");
         }
 
-        public async Task<bool> EditUserBalance(User editUser) //Edits balance of user, get current calues with GET and send new ones with POST. This is meant for when user sell/buy stock
+        public async Task<ActionResult> EditUserBalance(User editUser) //Edits balance of user, get current calues with GET and send new ones with POST. This is meant for when user sell/buy stock
         {
-            return await _db.EditUserBalance(editUser);
+            if (ModelState.IsValid)
+            {
+                bool ok = await _db.EditUserBalance(editUser);
+                if (!ok)
+                {
+                    _log.LogError("UserController: Could not edit userbal");
+                    return BadRequest("Could not edit userbal");
+                }
+                _log.LogInformation("UserController: User edited userbal");
+                return Ok("User edited userbal");
+            }
+            _log.LogError("UserController: Fault in InputVal ");
+            return BadRequest("Fault in InputVal");
 
         }
     }
